+++ eth/fetcher/tx_fetcher.go
@@ -26,7 +26,7 @@ import (
 	mapset "github.com/deckarep/golang-set"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/mclock"
-	"github.com/ethereum/go-ethereum/core"
+	"github.com/ethereum/go-ethereum/core/txpool"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/log"
 	"github.com/ethereum/go-ethereum/metrics"
@@ -281,7 +281,7 @@ func (f *TxFetcher) Enqueue(peer string, txs []*types.Transaction, direct bool) 
 			// Track the transaction hash if the price is too low for us.
 			// Avoid re-request this transaction when we receive another
 			// announcement.
-			if err == core.ErrUnderpriced || err == core.ErrReplaceUnderpriced {
+			if err == txpool.ErrUnderpriced || err == txpool.ErrReplaceUnderpriced {
 				for f.underpriced.Cardinality() >= maxTxUnderpricedSetSize {
 					f.underpriced.Pop()
 				}
@@ -291,10 +291,10 @@ func (f *TxFetcher) Enqueue(peer string, txs []*types.Transaction, direct bool) 
 			switch err {
 			case nil: // Noop, but need to handle to not count these
 
-			case core.ErrAlreadyKnown:
+			case txpool.ErrAlreadyKnown:
 				duplicate++
 
-			case core.ErrUnderpriced, core.ErrReplaceUnderpriced:
+			case txpool.ErrUnderpriced, txpool.ErrReplaceUnderpriced:
 				underpriced++
 
 			default:
